PLAT-42981: Update pattern samples to reflect the current standards and way of doing prerendering
