+++ pattern-activity-panels-deep-linking/src/index.js
@@ -1,8 +1,24 @@
 import React from 'react';
 import {render} from 'react-dom';
-import App from './app';
+import {Provider} from 'react-redux';
+import App from './App';
+import configureStore from './store';
 
-render(
-	<App />,
-	document.getElementById('root')
+// set default launch path
+const launchParam = '/first/second';
+const store = configureStore({
+	path: launchParam
+});
+
+const appElement = (
+	<Provider store={store}>
+		<App />
+	</Provider>
 );
+
+// In a browser environment, render the app to the document.
+if (typeof window !== 'undefined') {
+	render(appElement, document.getElementById('root'));
+}
+
+export default appElement;
