+++ pattern-activity-panels/package.json
@@ -20,7 +20,7 @@
   "repository": "https://github.com/enyojs/enact-samples",
   "enact": {
     "title": "ActivityPanels sample",
-    "isomorphic": "src/iso.js",
+    "isomorphic": true,
     "ri": {
       "baseSize": 24
     }
