+++ pattern-activity-panels/src/index.js
@@ -2,7 +2,11 @@ import React from 'react';
 import {render} from 'react-dom';
 import App from './App';
 
-render(
-	<App />,
-	document.getElementById('root')
-);
+const appElement = (<App />);
+
+// In a browser environment, render the app to the document.
+if (typeof window !== 'undefined') {
+	render(appElement, document.getElementById('root'));
+}
+
+export default appElement;
