+++ pattern-expandablelist-object/src/index.js
@@ -2,14 +2,11 @@ import React from 'react';
 import {render} from 'react-dom';
 import App from './App';
 
-let appElement = (<App />);
+const appElement = (<App />);
 
-// In a browser environment, render instead of exporting
+// In a browser environment, render the app to the document.
 if (typeof window !== 'undefined') {
-	render(
-		appElement,
-		document.getElementById('root')
-	);
+	render(appElement, document.getElementById('root'));
 }
 
 export default appElement;
