--- pattern-list-details-redux/src/iso.js
@@ -1,5 +1,0 @@
-import React from 'react';
-import App from './App';
-
-// Export the isomorphic entrypoint
-export default (<App />);
