+++ pattern-single-panel/package.json
@@ -20,7 +20,7 @@
   "repository": "https://github.com/enyojs/enact-samples",
   "enact": {
     "title": "Profile Photo Picker",
-    "isomorphic": "src/iso.js",
+    "isomorphic": true,
     "ri": {
       "baseSize": 24
     }
