+++ pattern-virtualgridlist-api/package.json
@@ -19,7 +19,7 @@
   "private": true,
   "repository": "https://github.com/enyojs/enact-samples",
   "enact": {
-    "isomorphic": "src/iso.js",
+    "isomorphic": true,
     "ri": {
       "baseSize": 24
     }
