+++ pattern-virtuallist-preserving-focus/src/index.js
@@ -1,24 +1,20 @@
-import {Provider} from 'react-redux';
 import React from 'react';
 import {render} from 'react-dom';
-
+import {Provider} from 'react-redux';
 import App from './App';
 import configureStore from './store';
 
 const store = configureStore();
 
-let appElement = (
+const appElement = (
 	<Provider store={store}>
 		<App />
 	</Provider>
 );
 
-// In a browser environment, render instead of exporting
+// In a browser environment, render the app to the document.
 if (typeof window !== 'undefined') {
-	render(
-		appElement,
-		document.getElementById('root')
-	);
+	render(appElement, document.getElementById('root'));
 }
 
 export default appElement;
