+++ tutorial-kitten-browser/src/index.js
@@ -1,9 +1,13 @@
 import React from 'react';
 import {render} from 'react-dom';
-
 import App from './App';
 
-render(
-	<App />,
-	document.getElementById('root') // provided by Enact's HTML template
-);
+const appElement = (<App />);
+
+// In a browser environment, render the app to the document.
+if (typeof window !== 'undefined') {
+	// 'root' div is provided by Enact's HTML template
+	render(appElement, document.getElementById('root'));
+}
+
+export default appElement;
